+++ packages/cubit/example/main.dart
@@ -1,9 +1,7 @@
 import 'package:cubit/cubit.dart';
 
 void main() async {
-  final cubit = CounterCubit()
-    ..listen(print)
-    ..increment();
+  final cubit = CounterCubit()..increment();
   await cubit.close();
 }
 
@@ -11,4 +9,10 @@ class CounterCubit extends Cubit<int> {
   CounterCubit() : super(initialState: 0);
 
   void increment() => emit(state + 1);
+
+  @override
+  void onTransition(Transition<int> transition) {
+    print(transition);
+    super.onTransition(transition);
+  }
 }