+++ packages/cubit/lib/cubit.dart
@@ -1,3 +1,5 @@
 library cubit;
 
 export 'src/cubit.dart';
+export 'src/cubit_stream.dart';
+export 'src/transition.dart';