+++ packages/cubit/test/cubits/counter_cubit.dart
@@ -1,8 +1,16 @@
 import 'package:cubit/cubit.dart';
 
 class CounterCubit extends Cubit<int> {
-  CounterCubit() : super(initialState: 0);
+  CounterCubit({this.onTransitionCallback}) : super(initialState: 0);
+
+  final void Function(Transition<int>) onTransitionCallback;
 
   void increment() => emit(state + 1);
   void decrement() => emit(state - 1);
+
+  @override
+  void onTransition(Transition<int> transition) {
+    onTransitionCallback?.call(transition);
+    super.onTransition(transition);
+  }
 }