+++ packages/cubit/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.7
+
+- feat: add `onTransition` to `cubit`
+- feat: `Cubit` extends `CubitStream`
+
 # 0.0.6
 
 - **BREAKING**: ignore duplicate states